requirements.txt
# Airflow (if running locally)
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-amazon

# Code formatting
black==23.7.0

# Core data processing
pandas==2.0.3

# Database connectivity
psycopg2-binary==2.9.7

# Development tools
pytest==7.4.0
python-dotenv==1.0.0

# Utility libraries
requests==2.31.0
sqlalchemy==2.0.20

duckdb
